requirements.txt
langchain 
langchain_pinecone 
langchain_groq 
langchain_google_genai
langchain_experimental
langchain_core
langgraph
langchain_huggingface
pypdf
streamlit-audiorecorder